♻️[Refactor]: View 파일명, 구조 정리 #374 #383
- 출시 스프린트 Backlog에 맞는 이름으로 변경
- Component로 활용되는 뷰 파일로 정리
