--- RelaxOn/App/Views/Components/BottomSoundPlayerBarView.swift
@@ -1,20 +1,0 @@
-//
-//  BottomSoundPlayerBarView.swift
-//  RelaxOn
-//
-//  Created by Doyeon on 2023/05/18.
-//
-
-import SwiftUI
-
-struct BottomSoundPlayerBarView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct BottomSoundPlayerBarView_Previews: PreviewProvider {
-    static var previews: some View {
-        BottomSoundPlayerBarView()
-    }
-}
