--- RelaxOn/App/Views/Components/SearchBarView.swift
@@ -1,20 +1,0 @@
-//
-//  SearchBarView.swift
-//  RelaxOn
-//
-//  Created by Doyeon on 2023/05/18.
-//
-
-import SwiftUI
-
-struct SearchBarView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct SearchBarView_Previews: PreviewProvider {
-    static var previews: some View {
-        SearchBarView()
-    }
-}
