+++ RelaxOn/App/Views/Components/SoundPlayerFullModalView.swift
@@ -1,8 +1,8 @@
 //
-//  SoundPlayerView.swift
+//  SoundPlayerFullModalView.swift
 //  RelaxOn
 //
-//  Created by 황석현 on 2023/04/19.
+//  Created by Doyeon on 2023/05/18.
 //
 
 import SwiftUI
@@ -10,8 +10,7 @@ import SwiftUI
 /**
  풀 모달 화면으로 보여지는 음원 플레이어 VIew
  */
-// TODO: FullModalSoundPlayerView로 네이밍 수정
-struct SoundPlayerView: View {
+struct SoundPlayerFullModalView: View {
     
     var body: some View {
         VStack {
@@ -67,8 +66,8 @@ struct SoundPlayerView: View {
     }
 }
 
-struct SoundPlayerView_Previews: PreviewProvider {
+struct FullModalSoundPlayerView_Previews: PreviewProvider {
     static var previews: some View {
-        SoundPlayerView()
+        SoundPlayerFullModalView()
     }
 }
