+++ RelaxOn/App/Views/Components/TimerProgressView.swift
@@ -45,7 +45,8 @@ struct TimerProgressView: View {
 
     var body: some View {
         ZStack {
-            CircleProgressBar(progress: $progress)
+            CircleProgressView(progress: $progress)
+            
             // TODO: View 파일 내에서는 계산식이 보이지 않도록 수정 필요
             Text(String(format: "%02d:%02d:%02d", max(remainingSeconds / 3600, 0), max((remainingSeconds % 3600) / 60, 0), max(remainingSeconds % 60, 0)))
                 .font(.system(size: 50))
@@ -61,31 +62,8 @@ struct TimerProgressView: View {
     }
 }
 
-/**
- 원형 프로그래스 바 View
- 타이머가 작동되고 있는 동안 남아있는 시간을 시각적으로 보여주는 기능
- */
-// TODO: CircleProgressBarView 로 네이밍 수정 -> 모델 객체로 보일 수 있음
-struct CircleProgressBar: View {
-
-    @Binding var progress: Double
-    @EnvironmentObject var timeData: TimerManager
-    @State var timer: Timer?
-
-    var body: some View {
-
-        ZStack {
-            Circle()
-                .stroke(lineWidth: 20.0)
-                .opacity(0.3)
-                .foregroundColor(.gray)
-
-            Circle()
-                .trim(from: 0.0, to: CGFloat(min(progress, 1.0)))
-                .stroke(style: StrokeStyle(lineWidth: 20.0, lineCap: .round, lineJoin: .round))
-                .foregroundColor(.black)
-                .rotationEffect(Angle(degrees: 270.0))
-
-        }
+struct TimerProgressView_Previews: PreviewProvider {
+    static var previews: some View {
+        TimerProgressView(timeData: TimerManager(), progress: .constant(0.5))
     }
 }
