+++ RelaxOn/App/Views/Listen/ListenListView.swift
@@ -41,110 +41,7 @@ struct ListenListView: View {
                 .onAppear {
                     viewModel.loadMixedSound()
                 }
-                PlayerBar()
-            }
-        }
-    }
-}
-
-/**
- 앱 내부 하단에 위치한 플레이어 바 형태의 View
- */
-// TODO: BottomSoundPlayerBarView 로 네이밍 수정
-struct PlayerBar: View {
-    
-    var body: some View {
-        
-        HStack {
-            Spacer(minLength: 80)
-
-            // TODO: 재생하는 사운드의 이미지 가져오기
-            Image(systemName: "play.fill")
-                .frame(width: 60, height: 60)
-                .background(.foreground.opacity(0.08)).cornerRadius(10)
-                .padding()
-            
-            Spacer()
-            
-            // TODO: 선택한 사운드 제목 가져오기
-            Text("Title")
-                .frame(width: 210,alignment: .leading)
-                .font(.title)
-            
-            Spacer()
-            
-            Button {
-                // TODO: 선택한 사운드를 재생
-            } label: {
-                Image(systemName: "play.fill")
-                    .frame(width: 60, height: 60)
-                    .foregroundColor(Color.black)
-            }
-            Spacer(minLength: 80)
-        }
-        .background(Color.systemGrey1)
-        .ignoresSafeArea()
-    }
-}
-
-/**
- 검색창 View
- */
-// TODO: SearchBarView 로 네이밍 수정 -> Model 객체처럼 보일 수 있음
-struct SearchBar: View {
-    // TODO: 검색기능 구현
-    @Binding var text: String
-
-    var body: some View {
-        HStack {
-            Image(systemName: "magnifyingglass")
-            TextField("저장한 나만의 소리를 검색해보세요", text: $text)
-                .foregroundColor(.primary)
-
-            if !text.isEmpty {
-                Button(action: {
-                    self.text = ""
-                }) {
-                    Image(systemName: "xmark.circle.fill")
-                        .foregroundColor(.primary)
-                }
-            }
-        }
-        .padding(EdgeInsets(top: 8, leading: 6, bottom: 8, trailing: 6))
-        .background(Color(.secondarySystemBackground))
-        .cornerRadius(10.0)
-    }
-}
-
-/**
- 커스텀 음원의 각 정보가 셀 안에 노출되는 View
- */
-struct ListenListCell: View {
-    // MARK: - Properties
-    
-    // TODO: 삭제 예정 -> viewModel로 관리
-    var title: String
-    var ImageName: String
-    
-    // TODO: 삭제 예정 -> enum으로 관리
-    var PlayButtonImageName: String = "play.fill"
-    var PauseButtonImageName: String = "pause.fill"
-    
-    var body: some View {
-        HStack {
-            Image(ImageName)
-                .frame(width: 60, height: 60)
-                .background(.foreground.opacity(0.08)).cornerRadius(10)
-            Text(title)
-                .font(.body)
-                .bold()
-            Spacer()
-            Button(action: {
-                // TODO: 재생 & 정지 토글 기능
-            }) {
-                Image(systemName: PauseButtonImageName)
-                    .frame(width: 32, height: 32)
-                    .foregroundColor(.black)
+                SoundPlayerBottomView()
             }
         }
     }
